+++ crates/miden-objects/src/account/delta/storage.rs
@@ -9,7 +9,7 @@ use super::{
     Word,
 };
 use crate::{
-    Digest, EMPTY_WORD, Felt, ONE, ZERO,
+    Digest, EMPTY_WORD, Felt, ZERO,
     account::{AccountStorage, StorageMap, StorageSlot},
 };
 // ACCOUNT STORAGE DELTA
@@ -116,19 +116,47 @@ impl AccountStorageDelta {
         self.values.iter().filter(|&(_, value)| value != &EMPTY_WORD)
     }
 
+    /// Appends the storage slots delta to the given `elements` from which the delta commitment will
+    /// be computed.
+    ///
     /// TODO: Make num_slots part of this struct.
     pub(super) fn append_delta_elements(&self, elements: &mut Vec<Felt>, num_slots: u8) {
+        const DOMAIN_VALUE: Felt = Felt::new(2);
+        const DOMAIN_MAP: Felt = Felt::new(3);
+
         for slot_idx in 0..num_slots {
-            match self.values().get(&slot_idx) {
-                Some(new_slot_value) => {
-                    // Append [was_slot_changed, 0, 0, 0].
-                    elements.extend_from_slice(&[ONE, ZERO, ZERO, ZERO]);
-                    elements.extend_from_slice(new_slot_value);
+            let slot_idx_felt = Felt::from(slot_idx);
+
+            // The storage delta ensures that the value slots and map slots do not have overlapping
+            // slot indices, so at most one of them will return `Some` for a given slot index.
+            match self.values.get(&slot_idx) {
+                Some(new_value) => {
+                    elements.extend_from_slice(&[DOMAIN_VALUE, slot_idx_felt, ZERO, ZERO]);
+                    elements.extend_from_slice(new_value);
                 },
                 None => {
-                    // TODO: Handle map slots and slots for which no delta is provided.
-                    elements.extend_from_slice(&EMPTY_WORD);
-                    elements.extend_from_slice(&EMPTY_WORD);
+                    if let Some(map_delta) = self.maps().get(&slot_idx) {
+                        if map_delta.is_empty() {
+                            continue;
+                        }
+
+                        for (key, value) in map_delta.leaves() {
+                            elements.extend_from_slice(key.as_elements());
+                            elements.extend_from_slice(value);
+                        }
+
+                        let num_changed_entries = Felt::try_from(map_delta.num_entries()).expect(
+                            "number of changed entries should not exceed max representable felt",
+                        );
+
+                        elements.extend_from_slice(&[
+                            DOMAIN_MAP,
+                            slot_idx_felt,
+                            num_changed_entries,
+                            ZERO,
+                        ]);
+                        elements.extend_from_slice(&EMPTY_WORD);
+                    }
                 },
             }
         }
@@ -247,6 +275,11 @@ impl StorageMapDelta {
     /// Creates a new storage map delta from the provided leaves.
     pub fn new(map: BTreeMap<Digest, Word>) -> Self {
         Self(map)
+    }
+
+    /// Returns the number of changed entries in this map delta.
+    pub fn num_entries(&self) -> usize {
+        self.0.len()
     }
 
     /// Returns a reference to the updated leaves in this storage map delta.